\name{concoR}
\alias{concoR}

\title{The concoR Package}

\description{
concoR is a package built to allow use the use of the CONCOR (CONvergence of iterated CORrelations) algorithm in R. This package includes the \code{\link{concor}} algorithm itself, and some useful functions for plotting and interpreting the outputs. The possible plotting methods included are plotting the overall network with with the \code{\link{concor}} partitions as colors (using the \code{igraph} package's framework), plotting a blockmodel formed from the \code{\link{concor}} partitions (using the \code{sna} package's framework), and plotting reduced network graphs showing the connections between \code{\link{concor}} partitions (again using the \code{igraph} package's framework to plot).
}

\author{
Tyme Suda \email{suda.4@wright.edu}

Adrienne Traxler \email{adrienne.traxler@wright.edu}
}

