\name{concoR}
\alias{concoR}

\title{The concoR Package}

\description{
concoR is a package built to allow use the use of the CONCOR (CONvergence of iterated
CORrelations) algorithm in R. This package includes the \code{\link{concor}} algorithm
itself, and some useful functions for plotting and interpreting the outputs. The plotting
methods included are: plotting the overall network with the \code{\link{concor}}
partitions as colors (using the \pkg{igraph} package), plotting a blockmodel using the
\code{\link{concor}} partitions (using the \pkg{sna} package), and plotting reduced
network graphs showing the connections between \code{\link{concor}} partitions (again
using \pkg{igraph}).
}

\author{
Tyme Suda \email{suda.4@wright.edu}

Adrienne Traxler \email{adrienne.traxler@wright.edu}
}

